update backend cleanup
